requirements.txt
natsort==8.4.0
numpy==1.24.3
opencv-python==4.5.3.56
Pillow==10.0.0
pyserial==3.5
kivy==2.2.1
